+++ ProWin/Sources/Engine/AudioBridge.h
@@ -1,0 +1,14 @@
+#ifndef AudioBridge_h
+#define AudioBridge_h
+
+#import <Foundation/Foundation.h>
+
+@interface AudioBridge : NSObject
+
++ (instancetype)sharedInstance;
+- (void)playSound;
+- (void)stopSound;
+
+@end
+
+#endif /* AudioBridge_h */